Originating revision: r479 | nedim_fresko | 2006-11-27 10:58:00 -0800 (Mon, 27 Nov 2006) | 27 lines
Sync up jump and cdc.
                                                                                
With these changes, CVM_MTASK roughly becomes the fork() function and
the mtask server for jump, and CVM_INCLUDE_JUMP becomes the app mgmt
framework. The framework should be able to exist on windows without
the mtask server.
                                                                                
The makefiles are changed to
1) Allow mtask builds only for linux and solaris, but not for other os'es.
2) Move mtask intelligence from shared to linux and solaris directories.
3) Make mtask=true if and only if include_jump=true for linux and solaris.
4) Allow CVM_INCLUDE_JUMP=true for windows, even if CVM_MTASK=false.
5) Change some included classes to be conditional on jump, not mtask,
since that is where they belong.
                                                                                
I am also putting back the files in the jump cdc-overlay directory.
                                                                                
There are still some packages named mtask in the cdc directory; those
will be slowly phased out in favor of jump. My first priority was to
be able to put back the mtask server with the new jump messaging, and
also to accommodate the cdc toolkit guys who want to use this stuff on
windows. So try to ignore some of the remaining mtask references which
will eventually migrate to the right names.
                                                                                
Reviewed for OSS correctness by Hinkmond.



git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/branches/resyncs@494 6dfe35fe-931f-0410-af5f-a91b034811cd