1. src/share/classes/com/sun/cdc/io/InternalConnector.java Added a new interface with a single open() method.
2. src/share/classes/com/sun/cdc/io/InternalConnectorImpl.java
This class implements InternalConnector interface and implements the open() method. It also has two protected methods:
getClassRoot() - sets the class root to com.sun.cdc.io (This is to load the correct protocol handler class for CDC)and
getProtocolClassLoader()- returns this class' classloader.


3. src/share/classes/sun/misc/MIDPInternalConnectorImpl.java
This class extends from InternalConnectorImpl and overrides two methods:
getClassRoot() - sets the class root to com.sun.midp.io and 
getProtocolClassLoader() - Loads the MIDP implementation classloader


4. src/share/classes/cldc/javax/microedition/io/Connector.java
Changed the static initializer of the class to read the property microedition.profiles. If this is set to MIDP, then load the class MIDPInternalConnectorImpl which is the implementation of InternalConnector interface.The open() method invokes the InternalConnector::open()

5. build/share/defs_cdc.mk
Added MIDPInternalConnectorImpl to CLASSLIB_CLASSES for CVM_DUAL_STACK

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@1978 6dfe35fe-931f-0410-af5f-a91b034811cd