- Fix undefined CVMCCMARMintrinsic_java_lang_Object_hashCodeGlue, CVMCCMARMintrinsic_java_lang_String_hashCodeGlue, and CVMmicrolockLockImpl symbol problem for Symbian build. They only exist if CVM_JIT_CCM_USE_C_SYNC_HELPER is disabled.
- Fix ARM RVCT assembler errors with the 'res_entry' macro in ccmglue_cpu.S.
Use a local variable so the ARM cpp will not add space when expending string 
concatenation.

- Rename sections included by ccmcodecachecopy_cpu.S in alphabetical
order. This is because some linker such as the ARM RVCT linker sorts
sections by attributes and section name. By doing this we can make sure
the CCM copied code are in the same order as they are included in 
ccmcodecachecopy_cpu.S.

Reviewed by Dean. 


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@2657 6dfe35fe-931f-0410-af5f-a91b034811cd