6569764: CDC backtraces do not include inlined compiled methods 6569765: CDC backtraces don't correctly identify compiled frames
Fixed CVMgcUnsafeFillInStackTrace() so it now looks at inlining info so it can include inlined frames in the backtrace.

Removed gcsafe checkpoint from CVMgcUnsafeFillInStackTrace(). It was a relic from the early days when we thought we wanted super short unsafe periods.

Fixed long standing bug in CVMgetStackTraceElement. The isCompiled java array was being treated as an an array of Refs rather than an array of Booleans. Thus indexing treated each cell as 32-bits instead of 8-bits, causing wrong values to be returned when indexing the array. This caused some methods in the backtrace to appear to be compiled when they were not, or not compiled when they were compiled. Also fixed how we reference the pcArray, which was being done as Refs instead of Ints, although this caused no problems since both are 32-bits.

Reviewed by Mark.

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@6061 6dfe35fe-931f-0410-af5f-a91b034811cd