Provide following new API in MIDPConfig for creating a MIDletClassLoader:
public static MIDletClassLoader
newMIDletClassLoader(String midpPath[], boolean enableFilter,
                     ClassLoader auxClassLoader)

This method can be used to specify if API hiding is enabled
for the specific MIDletClassLoader. If 'enableFilter' is false, 
API hiding will be disabled and midlet loaded by the 
MIDletClassLoader can access all CDC APIs. If 'auxClassLoader'
is set, it can be used to load class if the MIDletClassLoader 
and its parent classloaders fail to load a requested class.

Also remove the getMIDletClassLoader() method and static 
'midletCL' field from MIDPConfig.

Reviewed by Steve.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@6739 6dfe35fe-931f-0410-af5f-a91b034811cd