Fixed the following warning:
c:/home/cjp/phoneme/trunk/cdc/src/win32/native/java/io/Win32FileSystem_md.c(209) : warning C4293: '<<' : shift count negative or too big, undefined behavior

This was actually a bug. I looked at the generated code, and it resulted in truncation of the upper 32-bits of the 64-bit result this method returns. However, a test case run on the emulator showed that for any file 4gb and larger, the call to GetFileSize is returning 0x00000000ffffffff, so in the end fixing this bug did not make a difference.

Reviewed by Dean.
Did test builds on win32-arm-wm5 and ran test case on large file. Also inspected the generated code before and after the fix.


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@7524 6dfe35fe-931f-0410-af5f-a91b034811cd