Fixes for building with CVM_PRELOAD_LIB=false CVM_STATICLINK_LIBS=true, allowing for a smaller binary size on wm5.
In most cases, I just had to replace references to CVM_PRELOAD_LIB with CVM_STATICLINK_LIBS. In some cases part of the CVM_PRELOAD_LIB logic was correct, and part had to be moved to a separate CVM_STATICLINK_LIBS section.

runMidlet.gmk changes were a bit more complex. There were two versions of the same rule, one for CVM_PRELOAD_LIB=true and one for CVM_PRELOAD_LIB=false. The rules were exactly the same except for the presence of the -Dsun.midp.library.name option in the rule command, so I merged them into one unconditional rule, and made the -Dsun.midp.library.name conditional.

In JavaAPILister, I made a change so it properly deals with not having any classes in the generated cre[], which was causing a compiler error. Instead cre[] is not generated, and the reference to it in CVMdualStackMemberFilter is changed to NULL.

Reviewed by Dean and Jiangli
Tested with numerous build flavors on wm5 and running TCK.

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@7847 6dfe35fe-931f-0410-af5f-a91b034811cd