6658980: cdc should support builds where part of the port is in a separate component
You can now place part of a CVM/CDC/FP port in a different component. The part that is placed in the separate component can include any of the following directories. They can even be spread over multiple components:

        build/<os>
        build/<cpu>
        build/<os>-<cpu>
        build/<os>-<cpu>-<device>
        src/<os>
        src/<cpu>
        src/<os>-<cpu>
        src/<os>-<cpu>-<device>

The GNUmakefile in the device build directory controls where the parts of the port are coming from by defining one or more of the following variables. Note that they all default to CDC_DIR if not set in the GNUmakefile:

        CDC_OS_COMPONENT_DIR
        CDC_CPU_COMPONENT_DIR
        CDC_OSCPU_COMPONENT_DIR
        CDC_DEVICE_COMPONENT_DIR

It is not possible, for example, to put the build/<os> directory in one component and the src/<os> directory in a different one, since they are both controlled by CDC_OS_COMPONENT_DIR. The following is what will typically go in a GNUmakefile for a port located in a different component:

        ABSPATH = $(shell mkdir -p $(1); cd $(1); echo `pwd`)
        ifndef COMPONENTS_DIR
        COMPONENTS_DIR  := $(call ABSPATH, ../../..)
        endif
        CDC_COM_DIR     := $(call ABSPATH, ../..)
        ifndef CDC_DIR
        CDC_DIR         := $(COMPONENTS_DIR)/cdc
        endif

        CDC_OS_COMPONENT_DIR      = $(CDC_COM_DIR)
        CDC_CPU_COMPONENT_DIR     = $(CDC_COM_DIR)
        CDC_OSCPU_COMPONENT_DIR   = $(CDC_COM_DIR)
        CDC_DEVICE_COMPONENT_DIR  = $(CDC_COM_DIR)

        include $(CDC_DIR)/build/share/top.mk

Many path references in the makefiles had to be changed to be absolute instead of relative. This is because some makefiles are now located in different components and must be accessed with absolute paths. Also, since the device build directory can be in a separate component now, using relative paths to find cdc component files no longer works. The following are all common change patterns needed in the makefiles to get the cdc makefiles to look in the proper place for files:

        .. ==> $(CDC_DIR)/build
        .. ==> $(CDC_OS_COMPONENT_DIR)/build
        .. ==> $(CDC_CPU_COMPONENT_DIR)/build
        .. ==> $(CDC_OSCPU_COMPONENT_DIR)/build
        .. ==> $(CDC_DEVICE_COMPONENT_DIR)/build
        $(CVM_TOP) ==> $(CDC_DIR)
        $(CVM_TOP) ==> $(CDC_OS_COMPONENT_DIR)
        $(CVM_TOP) ==> $(CDC_CPU_COMPONENT_DIR)
        $(CVM_TOP) ==> $(CDC_OSCPU_COMPONENT_DIR)
        $(CVM_TOP) ==> $(CDC_DEVICE_COMPONENT_DIR)

Choosing which of the above to use depends on whether "share", $(TARGET_OS), $(TARGET_CPU), $(TARGET_OS)-$(TARGET_CPU), or $(TARGET_OS)-$(TARGET_CPU)-$(TARGET_DEVICE) is used in the path.

Since most paths are now absolute instead of relative, it was necessary to add POSIX2HOST calls in more places.

There was a bug in the $(CVM_BUILD_DEFS_H) rule when using CVM_BUILD_SUBDIR_NAME, because it referenced ${CVM_BUILD_TOP}/../share, which would not exist if you are in a subdir.  It now uses the absolute $(CDC_DIR)/build/share path instead.

Tested with wm5, win32/x86 and linux/x86 builds, and also tested by moving wm5, win32/x86, and linux/x86 builds to another component and building from there.

Reviewed by Dean and Jiangli

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@9524 6dfe35fe-931f-0410-af5f-a91b034811cd