Change #1: Slightly faster builds on win32 (about 15% for CDC-HI and 7% when MIDP and all the JSRs are included). This was done by getting rid of the need for some cygpath calls that were made on every compilation (we now have 1 instead of 3)
Change #2: Fixed source file path names stored with the debugging symbols so they are always relative. At one point they used to all be relative to the build directory. This is what you want, because you can just point the debugger to the build directory (even in a copy of the sources) and it will find all the files. When other components (JSRs) starting getting integrated, they were compiled using absolute paths. After the recent changes to allow parts of CDC ports to be in external components, now all the files compile with absolute paths. If the debugging info has an absolute path, and the path is not valid on the target, then the debugger ignores the path, and you end up having to point the debugger to each directory containing a source file.

Details:

There is a new abs2rel macro that takes care of the conversion from and absolute to a relative path, so what follows are mostly descriptions of where this is now done.

At first I tried changing things like the vpath paths, but then there were problems with the vpaths being relative while the rule dependencies that triggered the vpath rules were relative. gnumake does not canonicalize paths to see if they match, so this resulted in gnumake not being able to find rules for targets.

The easier approach proved to be just use abs2rel instead of POSIX2HOST on the source file argument when compiling. While making this change I was also able to consolidate the shared and win32 CC_CMD_XXX macros, so win32 only needs to specify compileCC now.

I made the object files relative by using abs2rel on CVM_OBJDIR and other similar macros that point to object directories. This allowed me to remove the POSIX2HOST on the object file in the compileCC macro, which speeds up win32 builds.

I also made all of the -I arguments relative. This shortens the command lines, making them a bit easier to read in verbose mode. This means abs2rel is called instead of POSIX2HOST. However, this is now done just once rather than for every compilation, so this also speeds up win32 builds. This change got tricky since I needed to use := to force the evaluation of ALL_INCLUDE_FLAGS. I had to move this from defs.mk to rules.mk so it would not be done too soon. JVMTI makefiles made this harder since they have rule dependencies that add to CVM_INCLUDE_DIRS. By the time the rule dependency was executed, we no longer look at CVM_INCLUDE_DIRS, so instead I changed to make the JVMTI rules add to ALL_INCLUDE_FLAGS.

I moved the inclusion of tool.gmk to much earlier in defs.mk so abs2rel could be used with a := assignment in defs.mk. I would have moved it to top.mk, but it requires that the HOST_OS variable be setup, and I didn't want to end up pulling a bunch of stuff into top.mk for this.

I added "." to CVM_INCLUDE_DIRS. It is needed when compiling jitcodegen.c. It was just a fluke that not having this never caused problems before when a #include of "../../build/<buildir>/generated/javavm/runtime/..." was used. The only reason it worked is because it was relative to paths like "../../src/linux". This stopped working when I made some changes that caused the #include to become "generated/javavm/runtime/...".

Reviewed by Dean
Tested with numerous Linux and Win32 builds.

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@9670 6dfe35fe-931f-0410-af5f-a91b034811cd