Fixes for ARM .so support. We were generating some PC24 entries in our ASM glue, which caused problems on the netwinder. See the section 2.24 in the following:
http://www.netwinder.org/faq.html

On the omap board, it would link and load OK, but would crash.

The PC24 references were caused by an assembler function in one file branching to an assembler function in another file using the exported symbol. For some reason this resulted in a PC24 reference that wasn't fixed up until runtime, and the loader can't handle it. If you instead branch to a local label (which we can do since the files are all compiled together), then you still get a PC24 reference, but it is fixed up at link time. The other choice was to start using the CALL_VM_FUNTION macro, but this does an "ldr pc", which is quite a bit slower, and some of this code is performance critical. 

Reviewed by Dean


git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@16855 6dfe35fe-931f-0410-af5f-a91b034811cd