Fix %gp support on MIPS, especially when building as a .so.
The LA macro has moved from asmmacros_arch.h to asmmacros_cpu.h.

In CVMJITgoNative, %gp is saved to the ccee so it can be used by the CCM glue. The LA macro now always uses the "la" mnemonic instead of using ALU operations. LA now also always loads %gp from the ccee first, so "la" will work.

Since LA always loads %gp, there may occassionally be times when it is loaded unecessarily, but this shouldn't be in anything that is performance critical. However, if need be, we could switch some of the LA references to "la". Also, %gp is considered volatile by the JIT, so there is no need to restore before %gp returning to the JIT'd code.

invokenative_mips.S and jit_cpu.S now use "la" instead of LA because they don'twant to load %gp from the ccee (%gp is setup in a different way for non ccm code).

I moved the CVMJNIinvokeNative jump tables to the .data section so they will work when cvm is built as a .so, and you don't need to use .gpword to avoid relocations in the .text section.

Since CVMJNIinvokeNative now uses "la" instead of "LA", it now also sets up, stores, and restores %gp as needed.

I got rid of the memmove hack that called memmove through a pointer stored in the ccee. This use to be done because there was a problem calling memove directly when using the old LA macro that made use of ALU instructions. It tried to create a relocation in the .text section, which could only be resolved at runtime since memmove is in another executable (and you can't do runtime relocations in the .text section). Since we use "la" now to load the address of memmove, there's no longer a need to go through a pointer in the ccee.

Reviewed by Dean

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@16856 6dfe35fe-931f-0410-af5f-a91b034811cd