6818557: Assertion in split_verify.c
This is an edge case that is covered by both verifiers, but the split verifier is making assumptions with the constant pool. In the full verifier, it is covered by the code below the following comment

        /* The 2nd edition VM of the specification allows field
         * initializations before the superclass initializer,
         * if the field is defined within the current class.
         */

The problem is that this is the one case where the split verifier assumes that the cp entry is resolved, which it is not. The full verifier verifies the right class by just comparing ClassIDs. So I changed the split verifier to first compare ClassIDs, and if they are the same, then just use vClass to do the field lookup, rather than fetching the class from the constant pool.

Reviewed by Dean.

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@17328 6dfe35fe-931f-0410-af5f-a91b034811cd