Added the ClassLoader.noVerification flag to allow disabling of verification for a ClassLoader. It is only checked when -Xverify:remote (the default) is used. It is ignored for -Xverify:all and -Xverify:none.
Added the following API for setting the ClassLoader.noVerification flag:

        sun.misc.CVM.setNoVerification(ClassLoader cl, Boolean value);

The following can be used in your java main() to disable verification for any future loading of classes by the -classpath ClassLoader and the ExtClassLoader: 

        ClassLoader cl = ClassLoader.getSystemClassLoader();
        while (cl != null) {
            sun.misc.CVM.setNoVerification(cl, true);
            cl = cl.getParent();
        }

Replace CVM_NEEDS_VERIFY() with CVMloaderNeedsVerify(), since the loader now needs to be passed in to check ClassLoader.noVerification, plus it was too big to keep a macro.

CVM_NEEDS_VERIFY took one argument, which was the result if the verification level is REMOTE. It was always passed in as (loader != NULL) or !CVMisTrustedClassLoader. Now instead the "verifyTrusted" argument is passed to CVMloaderNeedsVerify(). If TRUE, which equates to (loader != NULL), trusted classloaders still need verification. If FALSE, CVMisTrustedClassLoader is called to see if verification is needed, but only if all other checks don't first produce a result.

Removed the CVM_NEEDS_VERIFY check in CVMsplitVerifyClass() since it had to have been made already before CVMsplitVerifyClass was called, plus it had a bug by passing in the loader to the macro rather than the boolean flag.

Reviewed by Dean

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@17560 6dfe35fe-931f-0410-af5f-a91b034811cd