Fix the following TCK failure:
vm/instr/putfield/putfield007/putfield00701m1/putfield00701m1.html

In opc_getfield and opc_putfield, "typeID", which is suppose to hold the typeID of the field, was getting clobbered when the typeID of the field's class was assigned to it. This broke the code at createResolveFBNode which expected "typeID" to hold the field typeID, not the class typeID. Fix came from ES with minor changes.

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/branches/typeid-extension-cjp@19233 6dfe35fe-931f-0410-af5f-a91b034811cd