During VM startup while we are calling CVMinitClass() on all the classes in classInitList, we cannot execute anything that relies on data retrieved via a ThreadLocal object. This is because ThreadLocal.get() relies on Thread.currentThread() returning a non-null value. It returns null because we have not constructed a Thread object yet.
This can be a problem if during class initialization an exception is thrown. Constructing the exception String will require ThreadLocal data if the excetion has a numeric float in it. The result will be throwing an NPE while trying to construct the exception String because Thread.currentThread() returned null.

Switching the order of thread initialization and class initialization fixes this problem. This gets the first Thread object created before there is any class initialization, so using ThreadLocal during class initialization will work.

git-svn-id: https://svn.java.net/svn/phoneme~svn/components/cdc/trunk@19295 6dfe35fe-931f-0410-af5f-a91b034811cd